classifier/requirements.txt
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
dnspython==2.6.1
google-api-core==2.18.0
google-api-python-client==2.125.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud-core==2.4.1
google-cloud-pubsub==2.21.0
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpc-google-iam-v1==0.13.0
grpcio==1.62.1
grpcio-status==1.62.1
httplib2==0.22.0
idna==3.6
joblib==1.3.2
numpy==1.26.4
oauthlib==3.2.2
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pymongo==4.6.3
pyparsing==3.1.2
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
scikit-learn==1.4.1.post1
scipy==1.13.0
threadpoolctl==3.4.0
uritemplate==4.1.1
urllib3==2.2.1
